do not remove the skeleton scripts if synchronizing another skeleton

+++ script/skeleton/synchronize.sh
@@ -40,7 +40,12 @@ cp -R script/* "${TARGET}/script"
 cp .gitignore "${TARGET}"
 cp Vagrantfile "${TARGET}"
 cd "${TARGET}" || exit 1
-rm -rf script/skeleton
+echo "${NAME}" | grep --quiet 'Skeleton$' && IS_SKELETON=true || IS_SKELETON=false
+
+if [ "${IS_SKELETON}" = false ]; then
+    rm -rf script/skeleton
+fi
+
 DASH=$(echo "${NAME}" | ${SED} --regexp-extended 's/([A-Za-z0-9])([A-Z])/\1-\2/g' | tr '[:upper:]' '[:lower:]')
 INITIALS=$(echo "${NAME}" | ${SED} 's/\([A-Z]\)[a-z]*/\1/g' | tr '[:upper:]' '[:lower:]')
 UNDERSCORE=$(echo "${DASH}" | ${SED} --regexp-extended 's/-/_/g')
